Update labels position, they now overlap matches
Previously the labels were put before the matches as prefixes.

When a match was after an indent, its label had an offset (equal to the
length of the indent).

When a match was at the beginning of a line, the labels already
overlapped the matches.

Ovelapping the matches seems to be the best solution.

As the new placement is more consistent with the placement done by other
plugins, it is the default placement from now on.

To use the old placement, set the `label_as_prefix` option to `true`.