feat: Add an option to reuse the last pattern
When the option `use_last_pattern` is set to `true`, the last pattern if
any, can be used for the next searches.

The goal of this feature is to avoid retyping the pattern.

But more importantly, combined with `max_pattern_length = 1` (last
commit), we can emulate a "repeat" command. where the user only have to
trigger a keymap and then only type one character for the label.

Here is an example of keymaps configuration :

```lua
-- normal behavior
vim.keymap.set("n", "s", sj.run)

-- "repeat" behavior
vim.keymap.set("n", "S", function()
	sj.run({
		max_pattern_length = 1,
		use_last_pattern = true,
	})
end)
```

**NOTICE** : For the last keymap, the options are passed "on demand" to
the main function and not set in the user configuration.