fix: issues with highlights when they are cleared
Most of the colorschemes use `highlight clear` which remove user
highlights and linked highlights might not be restored.

Starting from this commit, an `autocmd` is created to ensure that the
highlights are preserved even when switching colorschemes.

To use the old behavior, users have to add this to their configuration :
`use_highlights_autocmd = false`