feat: relative labels + forward/backward search
By default, labels are ordered from the top of the scope to the bottom,
and left to right, independent of the cursor position.

To have labels ordered depending on the cursor position, set
`relative_labels` option to `true`.

In order to have a better experience the `forward_search` option have
been added and can be used with `relative_labels`.