feat: Select a window
Users can now select the window that will be used for the search.

To do this, set `select_window` option to `true` when using `sj.run()`.

Users can also jump to a window without starting a search by using the
new function `sj.select_window()` and assigning it to a keymap.