feat: add the `stop_on_fail` option
Currently, the search stops when a pattern has no matches.
(The loop breaks and the highlights are cleared)

This behavior can mislead the user into thinking that there has been a jump.

To avoid this set the `stop_on_fail` option to `false`.
For compability, this option is set to `true` by default.